+++ src/PhpSpreadsheet/Reader/IReader.php
@@ -13,6 +13,8 @@ interface IReader
     public const SKIP_EMPTY_CELLS = 4;
     public const IGNORE_EMPTY_CELLS = 4;
 
+    public const IGNORE_ROWS_WITH_NO_CELLS = 8;
+
     public function __construct();
 
     /**